Adding fontawesome generated assets. -Not sure if I need to commit these but better safe than sorry
Last commit said to run a command to get assetic to dump files but it was wrong.
Command should be:
**
php app/console assets:install web
**
(no need to symlink the generated assets as it causes problems with svn)
You also might not need to run this command since I'm committing the generated files.