Moved creating a post reply action from the ThreadController to the PostController where it should live.
Added a 'mediaIncluded' field to a Thread and you can hardcode a value for a media file into the db for testing.

Fixed an invalid mapping for User and threads.

Added some transformers to transform thread to a number and media to a string (unused right now).

Changed basetemplate css so you can override a sidebarlinks block to make a different link active other than 'home' all the time.