*schema update required*
Fixed some messaging bugs caused by the change to the utility user logged in function.

Fixed a bug where you could view anyone's messages by manipulating the URL.

Fixed a bug where you would delete a message and it wouldn't be deleted.

Fixed a bug where you would archive a message and it wouldn't archive.

Styled the pagination of forums a wee bit.

Cleaned up and removed some old code that was left around and not functional anymore, particularly to do with messaging.