*composer update required*
Added the DoctrineFixtures bundle to enable loading of controlled set of data into database. Currently used to create a 'noreply' user and replaces the old method. Possibility to create test data to load into the dB in the future.

Re-wrote the code for displaying public member directory using knppaginator. Still not completely finished but a good start.

Removed a mistyped variablename and associated getter/setter methods for the UserProfile. Some restyling of code to put associated variables near each other.

Added a field for 'interestedInMentoringMentor' in the show and edit Profile views.