*schema update required*
When creating a newpost, the 'lastpostat' field is now set to the current time the post was created. This is different behaviour than previous.

Changed sorting method of getting top level threads for forum in the forumRepository.

Disabled changing the media file for a thread when you edit it. I don't think it makes sense to be able to change it, but you can change the title and the supplemental text.

Fix for permissions schema. Update to some wording with permission icons.